'kang'

+++ src/main/java/com/myspring/spring/qna/QnaService.java
@@ -144,14 +144,14 @@ public class QnaService {
 			QnaVO res = qnaMapper.getQna(qnaNo);
 			// 문의 삭제
 			int resQna = qnaMapper.deleteQna(qnaNo);
+			if (resQna == 0)
+				return new ResponseEntity<>(resQna, HttpStatus.INTERNAL_SERVER_ERROR);
 
 			// 댓글 삭제
 			if (res.isReply() == true) {
 				int resDelReply = qnaMapper.deleteReply(qnaNo);
 				if (resDelReply == 0)
 					return new ResponseEntity<>(resDelReply, HttpStatus.INTERNAL_SERVER_ERROR);
-				else
-					return new ResponseEntity<>(resDelReply, HttpStatus.OK);
 			} else {
 				// 답글이 삭제될때 원글의 reply가 false로 바꾸기
 				// reply글의 originalNo를 받아와서 그 originalNo의 reply를 false로 바꿔주기
@@ -159,14 +159,13 @@ public class QnaService {
 				if (res.getQnaNo() != res.getOriginalNo()) {
 					// qna가 삭제될때 reply도 같이 삭제하기
 					int resReply = qnaMapper.updateReplyFalse(res.getOriginalNo());
-					res.setOriginalNo(res.getQnaNo());
 					if (resReply == 0)
 						return new ResponseEntity<>(resReply, HttpStatus.INTERNAL_SERVER_ERROR);
 					else
 						return new ResponseEntity<>(resReply, HttpStatus.OK);
 				}
 			}
-		
+
 			File file;
 			File[] underDir;
 
@@ -174,14 +173,14 @@ public class QnaService {
 			underDir = file.listFiles();
 			for (int i = 0; i < underDir.length; i++) {
 				underDir[i].delete();
-			}	
+			}
 			entity = new ResponseEntity<>(HttpStatus.OK);
 		} catch (Exception e) {
 			e.printStackTrace();
 			entity = new ResponseEntity<>(HttpStatus.INTERNAL_SERVER_ERROR);
 		}
 		return entity;
-		
+
 //		
 //		QnaVO res = qnaMapper.getQna(qnaNo);
 //		// 문의 삭제
